[Ref] PAT: Silence -Wconversion warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18588 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/SampleFormats.cpp
@@ -967,7 +967,7 @@ static void PatchToSample(CSoundFile *that, SAMPLEINDEX nSample, GF1SampleHeader
 	sample.nLoopStart = sampleHeader.loopstart;
 	sample.nLoopEnd = sampleHeader.loopend;
 	sample.nC5Speed = sampleHeader.freq;
-	sample.nPan = (sampleHeader.balance * 256 + 8) / 15;
+	sample.nPan = static_cast<uint16>((sampleHeader.balance * 256 + 8) / 15);
 	if(sample.nPan > 256) sample.nPan = 128;
 	else sample.uFlags.set(CHN_PANNING);
 	sample.nVibType = VIB_SINE;
